pyproject.toml
[project]
name = "furby-project"
version = "0.1.0"
description = "Tools for extracting, editing, and rebuilding Furby ROM assets."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pillow",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
]

[project.scripts]
furby = "furby_tool.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
